66141: SystemML Engine Enhancements - fixing an incorrect parsing of ngroups in groupedAggregate

+++ SystemML/SystemML/DML/src/main/java/com/ibm/bi/dml/runtime/instructions/CPInstructions/ParameterizedBuiltinCPInstruction.java
@@ -129,7 +129,7 @@ public class ParameterizedBuiltinCPInstruction extends ComputationCPInstruction
 			
 			int ngroups = -1;
 			if ( params.get(Statement.GAGG_NUM_GROUPS) != null) {
-				ngroups = Integer.parseInt(params.get(Statement.GAGG_NUM_GROUPS));
+				ngroups = (int) Double.parseDouble(params.get(Statement.GAGG_NUM_GROUPS));
 			}
 			
 			// compute the result
